KAN-68
The most significant changes involve the addition of new classes and constructors, as well as the enhancement of existing classes with summary comments and new methods. The `FailedToCreateException<T>` class was added to handle exceptions when an instance creation fails. The `FailedToDeleteException<T>`, `FailedToUpdateException<T>`, `NotFoundException`, and `UserDoesNotExistException` classes were updated with summary comments and new constructors. The `CustomResponseMessage` class was enhanced with new methods to handle failed operations, and the `BaseDbModel` class was updated with a summary comment for the `Id` property.

1. The `FailedToUpdateException<T>` class was removed from the `FailedToCreateException - Copy.cs` file.
2. A new class `FailedToCreateException<T>` was added to the `FailedToCreateException.cs` file. This class handles exceptions when an instance creation of type `T` fails. It includes two constructors, one without parameters and one with an `id` parameter.
3. The `FailedToDeleteException<T>` class in the `FailedToDeleteException.cs` file was updated with summary comments. A new constructor was added that takes an `id` parameter.
4. The `FailedToUpdateException<T>` class in the `FailedToUpdateException.cs` file was updated. The class now handles exceptions when a failed update operation occurs. The constructors were updated to reflect this change.
5. The `NotFoundException` class in the `NotFoundException.cs` file was updated with summary comments. A new constructor was added that takes an `id` parameter.
6. The `UserDoesNotExistException` class in the `UserDoesNotExistException.cs` file was updated with summary comments.
7. The `CustomResponseMessage` class in the `ResponseMessage.cs` file was updated with summary comments. New methods were added to get string information of a failed creation, update, or deletion of an object.
8. The `BaseDbModel` class in the `BaseDbModel.cs` file was updated with a summary comment for the `Id` property.
