+++ Common/Models/BaseDbModel.cs
@@ -13,7 +13,9 @@ namespace Common.Models
     /// </summary>
     public class BaseDbModel : BaseModel
     {
-        [PrimaryKey("Id", false)]
+        /// <summary>
+        /// Get or sets the entity Id attribute
+        /// </summary>
         public int Id { get; set; }
     }
 }
